requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.1
greenlet==3.0.2
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
multidict==6.0.4
multipart==0.2.4
Pillow==10.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
yarl==1.9.4
